Final refactoring. Re-introduced and tested _focussedCallback
