+++ docs/_autosummary/pyxbmct.addoncontrols.rst
@@ -1,0 +1,35 @@
+pyxbmct.addoncontrols
+=====================
+
+.. automodule:: pyxbmct.addoncontrols
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+
+      Button
+      Edit
+      FadeLabel
+      Image
+      Label
+      List
+      RadioButton
+      Group
+      Skin
+      Slider
+      TextBox
+      ControlWithConnectCallback
+      ControlWithPlacedCallback
+      ControlWithRemovedCallback
+      ControlWithFocusedCallback
+   
+   
+
+   
+
