+++ docs/_autosummary/pyxbmct.addonwindow.rst
@@ -19,29 +19,11 @@ pyxbmct.addonwindow
       AddonWindow
       BlankDialogWindow
       BlankFullWindow
-      Button
-      Edit
-      FadeLabel
-      Image
-      Label
-      List
-      RadioButton
-      Group
-      Skin
-      Slider
-      TextBox
-      ControlWithConnectCallback
-      ControlWithPlacedCallback
-      ControlWithRemovedCallback
-   
-   
 
-   
-   
-   .. rubric:: Exceptions
 
-   .. autosummary::
-   
-      AddonWindowError
+
+
+
+
