+++ docs/_autosummary/pyxbmct.addonwindowerror.rst
@@ -1,0 +1,19 @@
+pyxbmct.addonwindowerror
+========================
+
+.. automodule:: pyxbmct.addonwindowerror
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Exceptions
+
+   .. autosummary::
+
+      AddonWindowError
+
+   
+
