+++ script.module.pyxbmct/lib/pyxbmct/__init__.py
@@ -10,7 +10,11 @@ Licence: GPL v.3 http://www.gnu.org/licenses/gpl.html
 
 from __future__ import absolute_import
 from .addonwindow import *
+from .abstractgrid import AbstractGrid
+from .addoncontrols import *
+from .addonwindowerror import AddonWindowError
 from .addonskin import BaseSkin
+from .kodiconstants import *
 
 __all__ = [
     'ALIGN_LEFT',
@@ -31,6 +35,7 @@ __all__ = [
     'ACTION_MOUSE_DRAG',
     'ACTION_MOUSE_MOVE',
     'ACTION_MOUSE_LEFT_CLICK',
+    'AbstractGrid',
     'AddonWindowError',
     'Label',
     'FadeLabel',
