+++ script.module.pyxbmct/lib/pyxbmct/addonwindow.py
@@ -7,12 +7,20 @@
 #
 # Licence: GPL v.3 <http://www.gnu.org/licenses/gpl.html>
 """
-This module contains all classes and constants of PyXBMCt framework
+This module contains all the Window classes of the PyXBMCt framework
 """
 
 from __future__ import absolute_import, division, unicode_literals
+import inspect
+from abc import ABCMeta, abstractmethod
+from .addonskin import skin
+from .abstractgrid import AbstractGrid
+from .addoncontrols import Button, List, RadioButton
+from .addonwindowerror import AddonWindowError
+from .kodiconstants import KEY_BUTTON_A, ACTION_PREVIOUS_MENU, ACTION_MOUSE_LEFT_CLICK
+from .xbmc4xbox import isXBMC4XBOX
 
-import platform
+_XBMC4XBOX = isXBMC4XBOX()
 
 # typing module only needed when running pytype and is not available on XBMC4XBOX AFAIK
 try:
@@ -20,895 +28,45 @@ try:
 except:
     pass
 
-XBMC4XBOX = platform.system() == "XBMC4Xbox"
-
 # kodi_six doesn't work on XBMC4XBOX
-if XBMC4XBOX:
+if _XBMC4XBOX:
     range = xrange
-    import xbmc, xbmcgui, inspect
+    import xbmcgui
 else:
+    from .addoncontrols import Slider, Edit
     from six.moves import range
-    from kodi_six import xbmc, xbmcgui
+    from kodi_six import xbmcgui
     from six import with_metaclass
 
-import os
-from abc import ABCMeta, abstractmethod
-from .addonskin import Skin
 
-skin = Skin()
-
-# Text alignment constants. Mixed variants are obtained by bit OR (|)
-ALIGN_LEFT = 0
-"""Align left"""
-ALIGN_RIGHT = 1
-"""Align right"""
-ALIGN_CENTER_X = 2
-"""Align center horisontally"""
-ALIGN_CENTER_Y = 4
-"""Align center vertically"""
-ALIGN_CENTER = 6
-"""Align center by both axis"""
-ALIGN_TRUNCATED = 8
-"""Align truncated"""
-ALIGN_JUSTIFY = 10
-"""Align justify"""
-
-# Kodi key action codes.
-# More codes available in xbmcgui module
-ACTION_PREVIOUS_MENU = 10
-"""ESC action"""
-ACTION_NAV_BACK = 92
-"""Backspace action"""
-ACTION_MOVE_LEFT = 1
-"""Left arrow key"""
-ACTION_MOVE_RIGHT = 2
-"""Right arrow key"""
-ACTION_MOVE_UP = 3
-"""Up arrow key"""
-ACTION_MOVE_DOWN = 4
-"""Down arrow key"""
-ACTION_MOUSE_WHEEL_UP = 104
-"""Mouse wheel up"""
-ACTION_MOUSE_WHEEL_DOWN = 105
-"""Mouse wheel down"""
-ACTION_MOUSE_DRAG = 106
-"""Mouse drag"""
-ACTION_MOUSE_MOVE = 107
-"""Mouse move"""
-ACTION_MOUSE_LEFT_CLICK = 100
-"""Mouse click"""
-
-KEY_BUTTON_A = 256
-"""XBMC4XBOX A button pressed"""
-
-
-def _set_textures(textures, kwargs):
-    """Set texture arguments for controls."""
-    for texture in textures:
-        if kwargs.get(texture) is None:
-            kwargs[texture] = textures[texture]
-
-
-class AddonWindowError(Exception):
-    """Custom exception"""
-    pass
-
-
-class ControlWithConnectCallback(object if XBMC4XBOX else with_metaclass(ABCMeta, object)):
-    """Abstract mixin class for controls that require a callback before events are connected to them"""
-
-    if XBMC4XBOX:
-        __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def _connectCallback(self, callback, window):
-        """
-        Called just before an event is connected to a control.
-        If true is returned callback is connected to the control.
-        If false is returned it is not.
-        If a new callback is returned then that is connected instead.
-        :param callback: the callback that is to be associated with the control
-        :param window: the window on which the connect method was called
-        """
-        # type: (typing.Callable, xbmcgui.Window) -> typing.Union[bool, typing.Callable]
-        raise NotImplementedError
-
-
-class ControlWithPlacedCallback(object if XBMC4XBOX else with_metaclass(ABCMeta, object)):
-    """Abstract mixin class for controls that require a callback after they have been placed"""
-
-    if XBMC4XBOX:
-        __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def _placedCallback(self, window, row, column, rowspan, columnspan, pad_x, pad_y):
-        """
-        Called after the control has been placed in a Window or Group.
-        :param window: the window in which the control has been placed
-        :param row: the row in which the control was placed (top left corner)
-        :param column: the column in which the control was placed (top left corner)
-        :param rowspan: the number of rows that the control takes up
-        :param columnspan: the number of columns that the control takes up
-        :param pad_x: the number of pixels of padding inside the cell around the left and right sides of the control
-        :param pad_y: the number of pixels of padding inside the cell around the top and bottom sides of the control
-        """
-        # type: (xbmcgui.Window, int, int, int, int, int, int) -> None
-        raise NotImplementedError
-
-
-class ControlWithRemovedCallback(object if XBMC4XBOX else with_metaclass(ABCMeta, object)):
-    """Abstract mixin class for controls that require a callback before they are removed"""
-
-    if XBMC4XBOX:
-        __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def _removedCallback(self, window):
-        """
-        Called just before a control is removed from the window
-        :param window: The window the control is being removed from
-        """
-        # type: (xbmcgui.Window) -> None
-        raise NotImplementedError
-
-
-class AbstractGrid(object if XBMC4XBOX else with_metaclass(ABCMeta, object)):
-    """
-    Grid functionality mixin.
-    
-
-    Mixin for parent widgets for other XBMC UI controls
-    much like Tkinter.Tk or PyQt QWidget class.
-    
-    .. warning:: This is an abstract class and is not supposed to be instantiated directly!
-    """
-    if XBMC4XBOX:
-        __metaclass__ = ABCMeta
-
-    @abstractmethod
-    def getRows(self):
-        """
-        Get grid rows count.
-        """
-        # type: () -> int
-        raise NotImplementedError
-
-    @abstractmethod
-    def getColumns(self):
-        """
-        Get grid columns count.
-        """
-        # type: () -> int
-        raise NotImplementedError
-
-    @abstractmethod
-    def getGridX(self):
-        # type: () -> int
-        raise NotImplementedError
-
-    @abstractmethod
-    def getGridY(self):
-        # type: () -> int
-        raise NotImplementedError
-
-    @abstractmethod
-    def getGridWidth(self):
-        # type: () -> int
-        raise NotImplementedError
-
-    @abstractmethod
-    def getGridHeight(self):
-        # type: () -> int
-        raise NotImplementedError
-
-    def getTileWidth(self):
-        if not hasattr(self, "_tile_width"):
-            self._tile_width = self.getGridWidth() // self.getColumns()
-        return self._tile_width
-
-    def getTileHeight(self):
-        if not hasattr(self, "_tile_height"):
-            self._tile_height = self.getGridHeight() // self.getRows()
-        return self._tile_height
-
-    @abstractmethod
-    def addControl(self, control):
-        # type: (xbmcgui.Control) -> None
-        raise NotImplementedError
-
-    @abstractmethod
-    def setAnimation(self, control):
-        # type: (xbmcgui.Control) -> None
-        raise NotImplementedError
-
-    @abstractmethod
-    def getWindow(self):
-        # type: () -> xbmcgui.Window
-        raise NotImplementedError
-
-    def placeControl(self, control, row, column, rowspan=1, columnspan=1, pad_x=5, pad_y=5):
-        """
-        Place a control within the window grid layout.
-
-        :param control: control instance to be placed in the grid.
-        :param row: row number where to place the control (starts from 0).
-        :param column: column number where to place the control (starts from 0).
-        :param rowspan: set when the control needs to occupy several rows.
-        :param columnspan: set when the control needs to occupy several columns.
-        :param pad_x: horisontal padding.
-        :param pad_y: vertical padding.
-        :raises: :class:`AddonWindowError` if a grid has not yet been set.
-
-        Use ``pad_x`` and ``pad_y`` to adjust control's aspect.
-        Negative padding values can be used to make a control overlap with grid cells next to it, if necessary.
-
-        Example::
-
-            self.placeControl(self.label, 0, 1)
-        """
-        # type: (xbmcgui.Control, int, int, int, int, int, int) -> None
-        tile_width = self.getTileWidth()
-        tile_height = self.getTileHeight()
-        control_x = (self.getGridX() + tile_width * column) + pad_x
-        control_y = (self.getGridY() + tile_height * row) + pad_y
-        control_width = tile_width * columnspan - 2 * pad_x
-        control_height = tile_height * rowspan - 2 * pad_y
-        control.setPosition(control_x, control_y)
-        control.setWidth(control_width)
-        control.setHeight(control_height)
-
-        self.addControl(control)
-        self.setAnimation(control)
-
-        if hasattr(control, "_placedCallback"):
-            control._placedCallback(self.getWindow(), row, column, rowspan, columnspan, pad_x, pad_y)
-
-
-# Inheritance for the sake of pytype seeing getX etc. xbmcgui.Control does not exist on XBMC4XBOX
-class ControlMixin(object if XBMC4XBOX else xbmcgui.Control):
-    """
-    Basic control functionality mixin.
-
-    Provides utility methods and wrappers for some existing methods to support PyXBMCt
-
-    .. warning:: This is a mixin class and is not supposed to be instantiated directly!
-    """
-
-    def __eq__(self, other):
-        if hasattr(other, 'getId'):
-            return self.getId() == other.getId()
-        return False
-
-    def isEnabled(self):
-        """
-        Determine if a control is enabled or not.
-
-        Example::
-
-            enabled = self.isEnabled()
-        """
-        # type: () -> bool
-        # Test this way so that a constructor is not needed
-        # to set the intial values
-        return not hasattr(self, "_is_enabled") or self._is_enabled
-
-    if XBMC4XBOX:
-        def isVisible(self):
-            """
-            Determine if the control is visible or not.
-
-            Example::
-
-                enabled = self.isVisible()
-            """
-            # type: () -> bool
-            return not hasattr(self, "_is_visible") or self._is_visible
-
-        def getX(self):
-            # type: () -> int
-            return self.getPosition()[0]
-
-        def getY(self):
-            # type: () -> int
-            return self.getPosition()[1]
-
-        def setVisible(self, is_visible):
-            """
-            Set whether the control is visible or not.
-
-            :param is_visible: boolean to determine if the control is visible.
-
-            Example::
-
-                self.setVisible(False)
-            """
-            # type: (bool) -> None
-            self._is_visible = is_visible
-            for ancestor in inspect.getmro(self.__class__):
-                if inspect.getmodule(ancestor) == xbmcgui:
-                    ancestor.setVisible(self, is_visible)
-                    return
-
-        def setEnabled(self, is_enabled):
-            """
-            Set whether the control is enabled or not.
-
-            :param is_enabled: boolean to determine if the control is enabled.
-
-            Example::
-
-                self.setEnabled(False)
-            """
-            # type: (bool) -> None
-            self._is_enabled = is_enabled
-            for ancestor in inspect.getmro(self.__class__):
-                if inspect.getmodule(ancestor) == xbmcgui:
-                    ancestor.setEnabled(self, is_enabled)
-                    return
-    else:
-        # isVisible is available by default so setVisible wrapper is not needed
-
-        def setEnabled(self, is_enabled):
-            """
-            Set whether the control is enabled or not.
-
-            :param is_enabled: boolean to determine if the control is enabled.
-
-            Example::
-
-                self.setEnabled(False)
-            """
-            # type: (bool) -> None
-            self._is_enabled = is_enabled
-            xbmcgui.Control.setEnabled(self, is_enabled)
-
-    def getMidpoint(self):
-        """
-        Get the (x,y) coordinates of the controls midpoint.
-
-        Example::
-
-            x, y = self.getMidpoint()
-        """
-        # type: () -> typing.Tuple[int, int]
-        x = self.getX()
-        y = self.getY()
-        width = self.getWidth()
-        height = self.getHeight()
-        return (x + (width / 2), y + (height / 2))
-
-
-class Label(ControlMixin, xbmcgui.ControlLabel):
-    """
-    Label(label, font=None, textColor=None, disabledColor=None, alignment=0,hasPath=False, angle=0)
-
-    ControlLabel class.
-
-    Implements a simple text label.
-
-    :param label: text string
-    :type label: str
-    :param font: font used for label text. (e.g. ``'font13'``)
-    :type font: str
-    :param textColor: hex color code of enabled label's label. (e.g. ``'0xFFFFFFFF'``)
-    :type textColor: str
-    :param disabledColor: hex color code of disabled label's label. (e.g. ``'0xFFFF3300'``)
-    :type disabledColor: str
-    :param alignment: alignment of label. **Note**: see ``xbfont.h``
-    :type alignment: int
-    :param hasPath: ``True`` = stores a path / ``False`` = no path.
-    :type hasPath: bool
-    :param angle: angle of control. (``+`` rotates CCW, ``-`` rotates CW)
-    :type angle: int
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    Example::
-
-        self.label = Label('Status', angle=45)
-    """
-
-    def __new__(cls, *args, **kwargs):
-        return super(Label, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class FadeLabel(ControlMixin, xbmcgui.ControlFadeLabel):
-    """
-    FadeLabel(font=None, textColor=None, _alignment=0)
-
-    Control that scrolls label text.
-
-    Implements a text label that can auto-scroll very long text.
-
-    :param font: font used for label text. (e.g. ``'font13'``)
-    :type font: str
-    :param textColor: hex color code of fadelabel's labels. (e.g. ``'0xFFFFFFFF'``)
-    :type textColor: str
-    :param _alignment: alignment of label. **Note**: see ``xbfont.h``
-    :type _alignment: int
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    Example::
-
-        self.fadelabel = FadeLabel(textColor='0xFFFFFFFF')
-    """
-
-    def __new__(cls, *args, **kwargs):
-        return super(FadeLabel, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class TextBox(ControlMixin, xbmcgui.ControlTextBox):
-    """
-    TextBox(font=None, textColor=None)
-
-    ControlTextBox class
-
-    Implements a box for displaying multi-line text.
-    Long text is truncated from below. Also supports auto-scrolling.
-
-    :param font: font used for text. (e.g. ``'font13'``)
-    :type font: str
-    :param textColor: hex color code of textbox's text. (e.g. ``'0xFFFFFFFF'``)
-    :type textColor: str
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    Example::
-
-        self.textbox = TextBox(textColor='0xFFFFFFFF')
-    """
-
-    def __new__(cls, *args, **kwargs):
-        return super(TextBox, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class Image(ControlMixin, xbmcgui.ControlImage):
-    """
-    Image(filename, aspectRatio=0, colorDiffuse=None)
-
-    ControlImage class.
-
-    Implements a box for displaying ``.jpg``, ``.png``, and ``.gif`` images.
-
-    :param filename: path or URL to an image file.
-    :type filename: str
-    :param aspectRatio: (values: ``0`` = stretch (default), ``1`` = scale up (crops), ``2`` = scale down (black bars)
-    :type aspectRatio: int
-    :param colorDiffuse: for example, ``'0xC0FF0000'`` (red tint)
-    :type colorDiffuse: str
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    Example::
-
-        self.image = Image('d:\images\picture.jpg', aspectRatio=2)
-    """
-
-    def __new__(cls, *args, **kwargs):
-        return super(Image, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class Button(ControlMixin, xbmcgui.ControlButton):
-    """
-    Button(label, focusTexture=None, noFocusTexture=None, textOffsetX=CONTROL_TEXT_OFFSET_X, textOffsetY=CONTROL_TEXT_OFFSET_Y, alignment=4, font=None, textColor=None, disabledColor=None, angle=0, shadowColor=None, focusedColor=None)
-
-    ControlButton class.
-
-    Implements a clickable button.
-
-    :param label: button caption
-    :type label: str
-    :param focusTexture: filename for focus texture.
-    :type focusTexture: str
-    :param noFocusTexture: filename for no focus texture.
-    :type noFocusTexture: str
-    :param textOffsetX: x offset of label.
-    :type textOffsetX: int
-    :param textOffsetY: y offset of label.
-    :type textOffsetY: int
-    :param alignment: alignment of label. **Note**: see ``xbfont.h``
-    :type alignment: int
-    :param font: font used for label text. (e.g. ``'font13'``)
-    :type font: str
-    :param textColor: hex color code of enabled button's label. (e.g. ``'0xFFFFFFFF'``)
-    :type textColor: str
-    :param disabledColor: hex color code of disabled button's label. (e.g. ``'0xFFFF3300'``)
-    :type disabledColor: str
-    :param angle: angle of control. (``+`` rotates CCW, ``-`` rotates CW)
-    :type angle: int
-    :param shadowColor: hex color code of button's label's shadow. (e.g. ``'0xFF000000'``)
-    :type shadowColor: str
-    :param focusedColor: hex color code of focused button's label. (e.g. ``'0xFF00FFFF'``)
-    :type focusedColor: str
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    Example::
-
-        self.button = Button('Status', font='font14')
-    """
-
-    def __new__(cls, *args, **kwargs):
-        textures = {'focusTexture': os.path.join(skin.images, 'Button', 'KeyboardKey.png'),
-                    'noFocusTexture': os.path.join(skin.images, 'Button', 'KeyboardKeyNF.png')}
-        _set_textures(textures, kwargs)
-        if kwargs.get('alignment') is None:
-            kwargs['alignment'] = ALIGN_CENTER
-        return super(Button, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class RadioButton(ControlMixin, xbmcgui.ControlRadioButton):
-    """
-    RadioButton(label, focusTexture=None, noFocusTexture=None, textOffsetX=None, textOffsetY=None, _alignment=None, font=None, textColor=None, disabledColor=None, angle=None, shadowColor=None, focusedColor=None, focusOnTexture=None, noFocusOnTexture=None, focusOffTexture=None, noFocusOffTexture=None)
-
-    ControlRadioButton class.
-
-    Implements a 2-state switch.
-
-    :param label: label text.
-    :type: str or unicode
-    :param focusTexture: filename for focus texture.
-    :type focusTexture: str
-    :param noFocusTexture: filename for no focus texture.
-    :type noFocusTexture: str
-    :param textOffsetX: x offset of label.
-    :type textOffsetX: int
-    :param textOffsetY: y offset of label.
-    :type textOffsetY: int
-    :param _alignment: alignment of label - *Note, see xbfont.h
-    :type _alignment: int
-    :param font: font used for label text. (e.g. 'font13')
-    :type font: str
-    :param textColor: hexstring -- color of enabled radio button's label. (e.g. '0xFFFFFFFF')
-    :type textColor: str
-    :param disabledColor: hexstring -- color of disabled radio button's label. (e.g. '0xFFFF3300')
-    :type disabledColor: str
-    :param angle: angle of control. (+ rotates CCW, - rotates CW)
-    :type angle: int
-    :param shadowColor: hexstring -- color of radio button's label's shadow. (e.g. '0xFF000000')
-    :type shadowColor: str
-    :param focusedColor: hexstring -- color of focused radio button's label. (e.g. '0xFF00FFFF')
-    :type focusedColor: str
-    :param focusOnTexture: filename for radio focused/checked texture.
-    :type focusOnTexture: str
-    :param noFocusOnTexture: filename for radio not focused/checked texture.
-    :type noFocusOnTexture: str
-    :param focusOffTexture: filename for radio focused/unchecked texture.
-    :type focusOffTexture: str
-    :param noFocusOffTexture: filename for radio not focused/unchecked texture.
-    :type noFocusOffTexture: str
-
-    .. note:: To customize RadioButton all 4 abovementioned textures need to be provided.
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    Example::
-
-        self.radiobutton = RadioButton('Status', font='font14')
-    """
-
-    def __new__(cls, *args, **kwargs):
-        if not XBMC4XBOX and xbmc.getInfoLabel('System.BuildVersion')[:2] >= '13':
-            textures = {'focusTexture': os.path.join(skin.images, 'RadioButton', 'MenuItemFO.png'),
-                        'noFocusTexture': os.path.join(skin.images, 'RadioButton', 'MenuItemNF.png'),
-                        'focusOnTexture': os.path.join(skin.images, 'RadioButton', 'radiobutton-focus.png'),
-                        'noFocusOnTexture': os.path.join(skin.images, 'RadioButton', 'radiobutton-focus.png'),
-                        'focusOffTexture': os.path.join(skin.images, 'RadioButton', 'radiobutton-nofocus.png'),
-                        'noFocusOffTexture': os.path.join(skin.images, 'RadioButton', 'radiobutton-nofocus.png')}
-        else:  # This is for compatibility with Frodo and earlier versions.
-            textures = {'focusTexture': os.path.join(skin.images, 'RadioButton', 'MenuItemFO.png'),
-                        'noFocusTexture': os.path.join(skin.images, 'RadioButton', 'MenuItemNF.png'),
-                        'TextureRadioFocus': os.path.join(skin.images, 'RadioButton', 'radiobutton-focus.png'),
-                        'TextureRadioNoFocus': os.path.join(skin.images, 'RadioButton', 'radiobutton-nofocus.png')}
-        _set_textures(textures, kwargs)
-        return super(RadioButton, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-# Edit is not supported on Xbox using the Python API
-if not XBMC4XBOX:
-    class Edit(ControlMixin, xbmcgui.ControlEdit):
-        """
-        Edit(label, font=None, textColor=None, disabledColor=None, _alignment=0, focusTexture=None, noFocusTexture=None, isPassword=False)
-
-        ControlEdit class. Edit is not supported on Xbox using the Python API
-
-        Implements a clickable text entry field with an on-screen keyboard.
-
-        :param label: text string.
-        :type label: str or unicode
-        :param font: [opt] font used for label text. (e.g. 'font13')
-        :type font: str
-        :param textColor: [opt] hexstring -- color of enabled label's label. (e.g. '0xFFFFFFFF')
-        :type textColor: str
-        :param disabledColor: [opt] hexstring -- color of disabled label's label. (e.g. '0xFFFF3300')
-        :type disabledColor: str
-        :param _alignment: [opt] lignment of label - *Note, see xbfont.h
-        :type _alignment: int
-        :param focusTexture: [opt] filename for focus texture.
-        :type focusTexture: str
-        :param noFocusTexture: [opt] filename for no focus texture.
-        :type noFocusTexture: str
-        :param isPassword: [opt] if ``True``, mask text value.
-        :type isPassword: bool
-
-        .. note:: You can use the above as keywords for arguments and skip certain optional arguments.
-            Once you use a keyword, all following arguments require the keyword.
-            After you create the control, you need to add it to the window with ``placeControl()``.
-
-        Example::
-
-            self.edit = Edit('Status')
-        """
-
-        def __new__(cls, *args, **kwargs):
-            textures = {'focusTexture': os.path.join(skin.images, 'Edit', 'button-focus.png'),
-                        'noFocusTexture': os.path.join(skin.images, 'Edit', 'black-back2.png')}
-            _set_textures(textures, kwargs)
-            return super(Edit, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class List(ControlMixin, xbmcgui.ControlList):
-    """
-    List(font=None, textColor=None, buttonTexture=None, buttonFocusTexture=None, selectedColor=None, _imageWidth=10, _imageHeight=10, _itemTextXOffset=10, _itemTextYOffset=2, _itemHeight=27, _space=2, _alignmentY=4)
-
-    ControlList class.
-
-    Implements a scrollable list of items.
-
-    :param font: string - font used for items label. (e.g. 'font13')
-    :param textColor: hexstring - color of items label. (e.g. '0xFFFFFFFF')
-    :param buttonTexture: string - filename for no focus texture.
-    :param buttonFocusTexture: string - filename for focus texture.
-    :param selectedColor: integer - x offset of label.
-    :param _imageWidth: integer - width of items icon or thumbnail.
-    :param _imageHeight: integer - height of items icon or thumbnail.
-    :param _itemTextXOffset: integer - x offset of items label.
-    :param _itemTextYOffset: integer - y offset of items label.
-    :param _itemHeight: integer - height of items.
-    :param _space: integer - space between items.
-    :param _alignmentY: integer - Y-axis alignment of items label - *Note, see xbfont.h
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    Example::
-
-        self.cList = List('font14', space=5)
-    """
-
-    def __new__(cls, *args, **kwargs):
-        textures = {'buttonTexture': os.path.join(skin.images, 'List', 'MenuItemNF.png'),
-                    'buttonFocusTexture': os.path.join(skin.images, 'List', 'MenuItemFO.png')}
-        _set_textures(textures, kwargs)
-        return super(List, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class Group(ControlMixin, xbmcgui.ControlGroup, AbstractGrid):
-    """
-    Group(rows, columns)
-
-    ControlGroup class.
-
-    Implements a secondary grid (with its own coordinate system) that controls can be placed in.
-    Allowing for finer control over where controls are placed in a window.
-
-    :param rows: integer - the number of rows in the grid
-    :param columns: integer - the number of columns in the grid
-
-    .. note:: After you create the control, you need to add it to the window with placeControl().
-
-    .. warning:: You must place Group in a window before adding controls to it!
-
-    Example::
-
-        self.group = Group(1,2)
-    """
-
-    def __new__(cls, rows, columns, *args, **kwargs):
-        # type: (int, int, typing.Any, typing.Any) -> None
-        return super(Group, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-    def __init__(self, rows, columns, *args, **kwargs):
-        # type: (int, int, typing.Any, typing.Any) -> None
-        self._rows = rows
-        self._columns = columns
-        self._controls = []
-        self._window = None
-
-    def getWindow(self):
-        # type: () -> xbmcgui.Window
-        if not self._window:
-            raise AddonWindowError("Group not placed!")
-        else:
-            return self._window
-
-    def getGridX(self):
-        # type: (None) -> int
-        return self.getX()
-
-    def getGridY(self):
-        # type: (None) -> int
-        return self.getY()
-
-    def getGridWidth(self):
-        # type: (None) -> int
-        return self.getWidth()
-
-    def getGridHeight(self):
-        # type: (None) -> int
-        return self.getHeight()
-
-    def getRows(self):
-        """
-        Get grid rows count.
-        """
-        # type: (None) -> int
-        return self._rows
-
-    def getColumns(self):
-        """
-        Get grid columns count.
-        """
-        # type: (None) -> int
-        return self._columns
-
-    def addControl(self, control):
-        # type: (xbmcgui.Control) -> None
-        # getWindow will throw an error if the Group is not placed
-        self.getWindow().addControl(control)
-        self._controls.append(control)
-
-    def setAnimation(self, control):
-        # type: (xbmcgui.Control) -> None
-        self._window.setAnimation(control)
-
-    def _removedCallback(self, window):
-        self.removeAllChildren()
-        self._window = None
-
-    def removeControl(self, control):
-        """
-        Remove a control from the window grid layout.
-
-        :param control: control instance to be removed from the grid.
-
-        Example::
-
-            self.removeControl(self.label)
-        """
-        # type: (xbmcgui.Control) -> None
-        # getWindow will throw an error if Group is not placed
-        self.getWindow().removeControl(control)
-        self._controls.remove(control)
-
-    def removeControls(self, controls):
-        """
-        Remove multiple controls from the window grid layout.
-
-        :param controls: an iterable of control instances to be removed from the grid.
-
-        Example::
-
-            self.removeControl(self.label)
-        """
-        # type: (typing.Iterable[xbmcgui.Control]) -> None
-        # Need to copy the list of controls as we are changing
-        # its size whilst iterating over it
-        for control in list(controls):
-            self.removeControl(control)
-
-    def removeAllChildren(self):
-        """
-        Removes all the Group's children (and all their children) from the window.
-
-        Example::
-
-            group.removeAllChildren()
-        """
-        # type: () -> None
-        self.removeControls(self._controls)
-
-    def setVisible(self, is_visible):
-        """
-        Sets the group (and it all its current children) to be either visible or invisible.
-
-        :param is_visible: determines whether or the group and its current children should be visible or not.
-
-        Example::
-
-            group.setVisible(False)
-        """
-        # type: (bool) -> None
-        xbmcgui.ControlGroup.setVisible(self, is_visible)
-        for control in self._controls:
-            control.setVisible(is_visible)
-
-    def setVisibleCondition(self, is_visible, allow_hidden_focus=False):
-        """
-        See the XBMC documentation
-        """
-        xbmcgui.ControlGroup.setVisibleCondition(self, is_visible, allow_hidden_focus)
-        for control in self._controls:
-            control.setVisibleCondition(is_visible, allow_hidden_focus)
-
-    def setEnabled(self, is_enabled):
-        """
-        Sets the group (and it all its current children) to be either enabled or disabled.
-
-        :param is_enabled: determines whether or the group and its current children should be enabled or not.
-
-        Example::
-
-            group.setVisible(setEnabled)
-        """
-        # type: (bool) -> None
-        xbmcgui.ControlGroup.setEnabled(self, is_enabled)
-        for control in self._controls:
-            control.setEnabled(is_enabled)
-
-    def setEnableCondition(self, is_enabled):
-        """
-        See the XBMC documentation
-        """
-        xbmcgui.ControlGroup.setEnableCondition(self, is_enabled)
-        for control in self._controls:
-            control.setEnableCondition(is_enabled)
-
-    def _placedCallback(self, window, *args, **kwargs):
-        """
-        Called once the grid has been placed
-        """
-        self._window = window
-
-
-# Slider is not supported on Xbox using the Python API
-if not XBMC4XBOX:
-    class Slider(ControlMixin, xbmcgui.ControlSlider):
-        """
-        Slider(textureback=None, texture=None, texturefocus=None, orientation=xbmcgui.HORIZONTAL)
-        
-        ControlSlider class.
-        
-        Implements a movable slider for adjusting some value.
-        
-        :param textureback: string -- image filename.
-        :param texture: string -- image filename.
-        :param texturefocus: string -- image filename.
-        :param orientation: int -- slider orientation
-        
-        .. note:: After you create the control, you need to add it to the window with placeControl().
-        
-        Example::
-        
-            self.slider = Slider()
-        """
-
-        def __new__(cls, *args, **kwargs):
-            textures = {'textureback': os.path.join(skin.images, 'Slider', 'osd_slider_bg.png'),
-                        'texture': os.path.join(skin.images, 'Slider', 'osd_slider_nibNF.png'),
-                        'texturefocus': os.path.join(skin.images, 'Slider', 'osd_slider_nib.png')}
-            _set_textures(textures, kwargs)
-            if xbmc.getInfoLabel('System.BuildVersion')[:2] >= '17':
-                kwargs['orientation'] = xbmcgui.HORIZONTAL
-            return super(Slider, cls).__new__(cls, -10, -10, 1, 1, *args, **kwargs)
-
-
-class AbstractWindow(AbstractGrid):
+class AbstractWindow(AbstractGrid if _XBMC4XBOX else with_metaclass(ABCMeta, AbstractGrid)):
     """
     Top-level control window.
-    
+
     The control windows serves as a parent widget for other XBMC UI controls
     much like Tkinter.Tk or PyQt QWidget class.
-    
+
     This class is a basic "skeleton" for a control window.
 
     .. warning:: This is an abstract class and is not supposed to be instantiated directly!
     """
 
+    if _XBMC4XBOX:
+        __metaclass__ = ABCMeta
+
     def __init__(self):
-        self.controls = []  # type: typing.List[xbmcgui.Control]
-        self.actions_connected = [] # type: typing.List[typing.Tuple[typing.Union[int, xbmcgui.Control], typing.List[typing.Callable]]]
-        self.controls_connected = [] # type: typing.List[typing.Tuple[typing.Union[int, xbmcgui.Control], typing.List[typing.Callable]]]
+        self._controls = []  # type: typing.List[xbmcgui.Control]
+        self._actions_connected = []  # type: typing.List[typing.Tuple[typing.Union[int, xbmcgui.Control], typing.List[typing.Callable]]]
+        self._controls_connected = []  # type: typing.List[typing.Tuple[typing.Union[int, xbmcgui.Control], typing.List[typing.Callable]]]
+        # Need access to this class and this is not easy to obtain
+        # as it is not the superclass of this one
+        self._XBMCWindowClass = None
+        for ancestor in inspect.getmro(self.__class__):
+            if inspect.getmodule(ancestor) == inspect.getmodule(xbmcgui):
+                self._XBMCWindowClass = ancestor
+            # break
+        if self._XBMCWindowClass is None:
+            raise AddonWindowError("Could not find Window parent class")
 
     def getWindow(self):
         return self
@@ -916,11 +74,11 @@ class AbstractWindow(AbstractGrid):
     def autoNavigation(self, vertical_wrap_around=True,
                        horizontal_wrap_around=True, include_disabled=False,
                        include_invisible=False, controls_subset=None,
-                       control_types=(Button, List, RadioButton) if XBMC4XBOX \
+                       control_types=(Button, List, RadioButton) if _XBMC4XBOX \
                                else (Button, List, RadioButton, Slider, Edit)):
         # type: (bool, bool, bool, bool, typing.Iterable[xbmcgui.Control], typing.Any) -> None
         if controls_subset is None:
-            controls = self.controls
+            controls = self._controls
         else:
             controls = controls_subset
 
@@ -1035,19 +193,19 @@ class AbstractWindow(AbstractGrid):
     def setGeometry(self, width_, height_, rows_, columns_, pos_x=-1, pos_y=-1):
         """
         Set width, height, Grid layout, and coordinates (optional) for a new control window.
-        
+
         :param width_: widgh of the created window.
         :param height_: height of the created window.
         :param rows_: # rows of the Grid layout to place controls on.
         :param columns_: # colums of the Grid layout to place controls on.
         :param pos_x: (opt) x coordinate of the top left corner of the window.
         :param pos_y: (opt) y coordinates of the top left corner of the window.
-        
+
         If pos_x and pos_y are not privided, the window will be placed
         at the center of the screen.
 
         Example::
-        
+
             self.setGeometry(400, 500, 5, 4)
         """
         # type: (int, int, int, int, int, int) -> None
@@ -1155,7 +313,7 @@ class AbstractWindow(AbstractGrid):
         # type: (typing.Union[int, xbmcgui.Control], typing.Callable) -> None
 
         if isinstance(event, int):
-            connect_list = self.actions_connected
+            connect_list = self._actions_connected
         else:  # Event is actually a control
             if hasattr(event, "_connectCallback"):
                 # Only connect the event if it returns true
@@ -1165,7 +323,7 @@ class AbstractWindow(AbstractGrid):
                     callback = should_connect
                 elif not should_connect:
                     return
-            connect_list = self.controls_connected
+            connect_list = self._controls_connected
         try:
             entry = next(entry for entry in connect_list if entry[0] == event)
 
@@ -1204,9 +362,9 @@ class AbstractWindow(AbstractGrid):
         """
         # type: (typing.Union[int, xbmcgui.Control], typing.Callable) -> None
         if isinstance(event, int):
-            event_list = self.actions_connected
+            event_list = self._actions_connected
         else:
-            event_list = self.controls_connected
+            event_list = self._controls_connected
 
         for event_index in range(len(event_list)):
             if event == event_list[event_index][0]:
@@ -1277,6 +435,121 @@ class AbstractWindow(AbstractGrid):
         """
         # type: (xbmcgui.Control) -> None
         pass
+
+    def setFocus(self, control):
+        # type: (xbmcgui.Control) -> None
+        do_set_focus = True
+        if hasattr(control, '_focusedCallback'):
+            do_set_focus = control._focusedCallback()
+        if do_set_focus:
+            self._XBMCWindowClass.setFocus(self, control)
+
+    def onFocus(self, control):
+        """
+        Catch focused controls.
+
+        :param control: is an instance of :class:`xbmcgui.Control` class.
+        """
+        # type: (xbmcgui.Control) -> None
+        if hasattr(control, '_focusedCallback'):
+            control._focusedCallback()
+
+    def addControl(self, control):
+        """
+        Wrapper for xbmcgui.Window.addControl.
+
+        :param control: the control to add.
+
+        .. note:: In most circumstances you should use placeControl.
+        .. note:: Only use this method if want to to place and element in a Group it using pixel coordinates
+
+        Example::
+
+            window.addControls(label)
+        """
+        # type: (xbmcgui.Control) -> None
+        self._controls.append(control)
+        self._XBMCWindowClass.addControl(self, control)
+
+    def addControls(self, controls):
+        """
+        Wrapper for xbmcgui.Window.addControls.
+
+        :param controls: iterable containing the controls to add.
+
+        .. note:: In most circumstances you should use placeControl.
+        .. note:: Only use this method if want to to place and element in a Group it using pixel coordinates
+
+        Example::
+
+            window.addControls([label, button])
+        """
+        # type: (typing.Iterable[xbmcgui.Control]) -> None
+        # addControls is not directly available on XBMC4XBOX
+        # so this implementation is the most portable
+        for control in controls:
+            self.addControl(control)
+
+    def removeControl(self, control):
+        """
+        Remove a control from the window grid layout.
+
+        :param control: control instance to be removed from the grid.
+
+        Example::
+
+            self.removeControl(self.label)
+        """
+        # type: (xbmcgui.Control) -> None
+        if hasattr(control, "_removedCallback"):
+            control._removedCallback(self)
+        self._XBMCWindowClass.removeControl(self, control)
+
+    def removeControls(self, controls):
+        """
+        Remove multiple controls from the window grid layout.
+
+        :param controls: an iterable of control instances to be removed from the grid.
+
+        Example::
+
+            self.removeControl(self.label)
+        """
+        # type: (typing.Iterable[xbmcgui.Control]) -> None
+        for control in controls:
+            self.removeControl(control)
+
+    def onAction(self, action):
+        """
+        Catch button actions.
+
+        :param action: an instance of :class:`xbmcgui.Action` class.
+        """
+        # type: (xbmcgui.Action) -> None
+        if action == ACTION_PREVIOUS_MENU:
+            self.close()  # pytype: disable=attribute-error
+        # On control is not called on XBMC4XBOX for subclasses of the built in
+        # controls. However onAction is.
+        elif _XBMC4XBOX and hasattr(action, 'getButtonCode') and action.getButtonCode() in (
+                KEY_BUTTON_A, ACTION_MOUSE_LEFT_CLICK):
+            control = self.getFocus()  # pytype: disable=attribute-error
+            if isinstance(control, (Button, RadioButton, List)) and control.isEnabled():
+                self.onControl(control)
+        else:
+            self._executeConnected(action, self._actions_connected)
+
+    def onControl(self, control):
+        """
+        Catch activated controls.
+
+        :param control: is an instance of :class:`xbmcgui.Control` class.
+        """
+        # type: (xbmcgui.Control) -> None
+        if (hasattr(self, 'window_close_button') and
+                control.getId() == self.window_close_button.getId()):
+            self.close()
+        else:
+            self._executeConnected(control, self._controls_connected)
 
 
 class AddonWindow(AbstractWindow):
@@ -1378,6 +651,7 @@ class AddonWindow(AbstractWindow):
         raise AddonWindowError('Window geometry is not defined! Call setGeometry first.')
 
     def getGridX(self):
+        # type: () -> int
         try:
             val = self.x + self.win_padding
         except AttributeError:
@@ -1385,6 +659,7 @@ class AddonWindow(AbstractWindow):
         return val + skin.x_margin
 
     def getGridY(self):
+        # type: () -> int
         try:
             val = self.y + self.win_padding
         except AttributeError:
@@ -1392,6 +667,7 @@ class AddonWindow(AbstractWindow):
         return val + skin.y_margin + skin.title_back_y_shift + skin.header_height
 
     def getGridWidth(self):
+        # type: () -> int
         try:
             val = self._width - 2 * self.win_padding
         except AttributeError:
@@ -1399,6 +675,7 @@ class AddonWindow(AbstractWindow):
         return val - 2 * skin.x_margin
 
     def getGridHeight(self):
+        # type: () -> int
         try:
             val = self._height - 2 * self.win_padding
         except AttributeError:
@@ -1416,102 +693,13 @@ class AddonWindow(AbstractWindow):
 
             self.setWindowTitle('My Cool Addon')
         """
+        # type: (str) -> None
         self.title_bar.setLabel(title)
 
     def getWindowTitle(self):
         """Get window title."""
+        # type: () -> str
         return self.title_bar.getLabel()
-
-    def addControl(self, control):
-        """
-        Wrapper for xbmcgui.Window.addControl.
-
-        :param control: the control to add.
-
-        .. note:: In most circumstances you should use placeControl.
-        .. note:: Only use this method if want to to place and element in a Group it using pixel coordinates
-
-        Example::
-
-            window.addControls(label)
-        """
-        self.controls.append(control)
-        xbmcgui.Window.addControl(self, control)
-
-    def addControls(self, controls):
-        """
-        Wrapper for xbmcgui.Window.addControls.
-
-        :param controls: iterable containing the controls to add.
-
-        .. note:: In most circumstances you should use placeControl.
-        .. note:: Only use this method if want to to place and element in a Group it using pixel coordinates
-
-        Example::
-
-            window.addControls([label, button])
-        """
-        # addControls is not directly available on XBMC4XBOX
-        # so this implementation is the most portable
-        for control in controls:
-            self.addControl(control)
-
-    def removeControl(self, control):
-        """
-        Remove a control from the window grid layout.
-
-        :param control: control instance to be removed from the grid.
-
-        Example::
-
-            self.removeControl(self.label)
-        """
-        if hasattr(control, "_removedCallback"):
-            control._removedCallback(self)
-        xbmcgui.Window.removeControl(self, control)
-
-    def removeControls(self, controls):
-        """
-        Remove multiple controls from the window grid layout.
-
-        :param controls: an iterable of control instances to be removed from the grid.
-
-        Example::
-
-            self.removeControl(self.label)
-        """
-        for control in controls:
-            self.removeControl(control)
-
-    def onAction(self, action):
-        """
-        Catch button actions.
-
-        :param action: an instance of :class:`xbmcgui.Action` class.
-        """
-        if action == ACTION_PREVIOUS_MENU:
-            self.close()
-        # On control is not called on XBMC4XBOX for subclasses of the built in
-        # controls. However onAction is.
-        elif XBMC4XBOX and hasattr(action, 'getButtonCode') and action.getButtonCode() in (
-                KEY_BUTTON_A, ACTION_MOUSE_LEFT_CLICK):
-            control = self.getFocus()
-            if isinstance(control, (Button, RadioButton, List)) and control.isEnabled():
-                self.onControl(control)
-        else:
-            self._executeConnected(action, self.actions_connected)
-
-    def onControl(self, control):
-        """
-        Catch activated controls.
-
-        :param control: is an instance of :class:`xbmcgui.Control` class.
-        """
-        if (hasattr(self, 'window_close_button') and
-                control.getId() == self.window_close_button.getId()):
-            self.close()
-        else:
-            self._executeConnected(control, self.controls_connected)
 
 
 class BlankFullWindow(AbstractWindow, xbmcgui.Window):
@@ -1563,6 +751,7 @@ class AddonFullWindow(AddonWindow, xbmcgui.Window):
         """
         Set the image for for the fullscreen background.
         """
+        # type: (str) -> None
         # Image for the fullscreen background.
         self.main_bg_img = skin.main_bg_img
         # Fullscreen background image control.
@@ -1580,6 +769,7 @@ class AddonFullWindow(AddonWindow, xbmcgui.Window):
 
             self.setBackground('/images/bacground.png')
         """
+        # type: (str) -> None
         self.main_bg.setImage(image)
