+++ script.module.pyxbmct/lib/pyxbmct/addonwindowerror.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# PyXBMCt framework module
+#
+# PyXBMCt is a mini-framework for creating Kodi (XBMC) Python addons
+# with arbitrary UI made of Controls - decendants of xbmcgui.Control class.
+# The framework uses image textures from Kodi Confluence skin.
+#
+# Licence: GPL v.3 <http://www.gnu.org/licenses/gpl.html>
+
+
+class AddonWindowError(Exception):
+    """Custom exception"""
+    pass
