+++ script.module.pyxbmct/lib/pyxbmct/xbmc4xbox.py
@@ -1,0 +1,4 @@
+import platform
+def isXBMC4XBOX():
+    # type: () -> bool
+    return platform.system() == "XBMC4Xbox"
